+++ Project/SCADA.Common/Messaging/Messages/Read.cs
@@ -42,7 +42,7 @@ namespace SCADA.Common.Messaging.Messages
             return request;
         }
 
-        public override Dictionary<Tuple<RegisterType, ushort>, BasePoint> PareseResponse(byte[] response)
+        public override Dictionary<Tuple<RegisterType, int>, BasePoint> PareseResponse(byte[] response)
         {
             throw new NotImplementedException();
         }
